moving com/alphawallet/ethereum/ERC721Token to org/tokenscript/attestation/ERC721Token and rename com/alphawallet/ethereum/ERC721TokenEth to com/alphawallet/ethereum/ERC721Token
