+++ src/main/java/com/alphawallet/ethereum/AttestationReturn.java
@@ -2,7 +2,7 @@ package com.alphawallet.ethereum;
 
 public class AttestationReturn
 {
-    public ERC721TokenEth[] ercToken;
+    public ERC721Token[] ercToken;
     public String identifier;
     public String ownerAddress;
     public String attestorAddress;
